back-end-tars/requirements.txt
# Core dependencies
requests>=2.31.0,<2.33.0
python-dotenv>=1.1.0,<1.2.0

# Azure AI Foundry SDK
azure-ai-projects==1.0.0b1
azure.identity==1.15.0

flask>=2.3.3,<3.0.0
bs4>=0.0.2

# API requirements with specific versions
# Updated FastAPI to a newer version that supports anyio>=4.5.0
fastapi>=0.109.0
uvicorn>=0.23.2
pydantic>=2.7.2,<3.0.0
httpx>=0.28.1,<0.29.0
python-multipart==0.0.6

# Updated anyio to be compatible with mcp
anyio>=4.5.0,<5.0.0